+++ apps/next-client/package.json
@@ -29,6 +29,7 @@
     "quill": "^2.0.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-error-boundary": "^5.0.0",
     "react-icons": "^5.4.0",
     "react-markdown": "^9.0.3",
     "react-quill": "^2.0.0",
